+++ src/doodle_core/metadata/shot.h
@@ -10,21 +10,21 @@ class DOODLE_CORE_API shot {
  public:
   enum class shot_ab_enum;
 
-  int64_t p_shot;
+  int32_t p_shot;
   shot_ab_enum p_shot_enum;
 
  public:
   std::string p_shot_ab;
   shot();
-  explicit shot(std::int64_t in_shot, shot_ab_enum in_ab);
-  explicit shot(std::int64_t in_shot, std::string in_ab);
+  explicit shot(std::int32_t in_shot, shot_ab_enum in_ab);
+  explicit shot(std::int32_t in_shot, std::string in_ab);
 
   // clang-format off
   enum class shot_ab_enum { None, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z };
   // clang-format on
 
-  [[nodiscard]] const std::int64_t &get_shot() const noexcept;
-  void set_shot(const std::int64_t &in_shot);
+  [[nodiscard]] const std::int32_t &get_shot() const noexcept;
+  void set_shot(const std::int32_t &in_shot);
 
   [[nodiscard]] std::string get_shot_ab() const noexcept;
   [[nodiscard]] shot_ab_enum get_shot_ab_enum() const noexcept;
@@ -63,12 +63,12 @@ namespace fmt {
  * @tparam
  */
 template <>
-struct formatter<::doodle::shot> : formatter<std::int64_t> {
+struct formatter<::doodle::shot> : formatter<std::int32_t> {
   template <typename FormatContext>
   auto format(const ::doodle::shot &in_, FormatContext &ctx) const -> decltype(ctx.out()) {
     format_to(ctx.out(), "sc_");
 
-    formatter<std::int64_t>::format(in_.p_shot, ctx);
+    formatter<std::int32_t>::format(in_.p_shot, ctx);
     if (in_.p_shot_enum != doodle::shot::shot_ab_enum::None)
       format_to(ctx.out(), magic_enum::enum_name(in_.p_shot_enum));
     return ctx.out();
